+++ src/parser/lexer/debug.h
@@ -10,29 +10,29 @@ char invert_brace_char(char brace);
 
 int8_t sprintf_token_slice(
     const struct Token tokens[],
-    usize ntokens,
+    uint16_t ntokens,
     char * output,
-    usize output_sz    
+    uint16_t output_sz    
 );
 
 int8_t sprintf_token_slice_by_ref(
     const struct Token *tokens[],
-    usize ntokens,
+    uint16_t ntokens,
     char * output,
-    usize output_sz    
+    uint16_t output_sz    
 );
 
 int8_t sprintf_lexicon_slice(
     const enum Lexicon tokens[],
-    usize ntokens,
+    uint16_t ntokens,
     char * output,
-    usize output_sz    
+    uint16_t output_sz    
 );
 
 int8_t sprint_src_code(
     char * output,
-    usize output_sz,
-    usize *nbytes,
+    uint16_t output_sz,
+    uint16_t *nbytes,
     const char * source,
     const struct Token *token
 );
